KP-7440 Move Metax request fixtures
