+++ tests/test_metax_api.py
@@ -4,45 +4,16 @@ import requests
 import requests_mock
 import metax_api
 
-@pytest.fixture
-def metax_base_url():
-    """Metax API"""
-    return "https://metax-service.fd-staging.csc.fi/v3"
-
-@pytest.fixture
-def kielipankki_datacatalog_id():
-    """Data catalog identifier for Kielipankki data."""
-    return "urn:nbn:fi:att:data-catalog-kielipankki-v3"
-
-def _get_file_as_string(filename):
-    """Return given file as string."""
-    with open(filename) as infile:
-        return infile.read()
-
-@pytest.fixture
-def dataset_pid():
-    """Return PID of sample record."""
-    return "urn.fi/urn:nbn:fi:lb-2016101210"
-
-@pytest.fixture
-def mock_get_response_json():
-    """A mocked response for a single dataset from Metax"""
-    return _get_file_as_string("tests/test_data/metax_single_record_response.json")
-
-@pytest.fixture
-def mock_requests_get_record(mock_get_response_json, metax_base_url, kielipankki_datacatalog_id, dataset_pid):
-    """A mock GET request to Metax."""
-    with requests_mock.Mocker() as mocker:
-        mocker.get(f"{metax_base_url}/datasets?data_catalog_id={kielipankki_datacatalog_id}&persistent_identifier={dataset_pid}", text=mock_get_response_json)
-        yield mocker
-
-def test_check_if_dataset_record_pid_in_datacatalog(dataset_pid, mock_requests_get_record, metax_base_url, kielipankki_datacatalog_id):
-    """Check that a dataset with a given PID exists in Metax by making a proper GET request to specified url."""
+def test_check_if_dataset_record_pid_in_datacatalog(
+        dataset_pid, mock_requests_get_record, metax_base_url, kielipankki_datacatalog_id):
+    """Check that a dataset with a given PID exists in Metax by making a proper GET request to 
+    specified url."""
     result = metax_api.check_if_dataset_record_in_datacatalog(dataset_pid)
     assert len(mock_requests_get_record.request_history) == 1
     assert mock_requests_get_record.request_history[0].method == "GET"
     assert mock_requests_get_record.request_history[0].url == f"{metax_base_url}/datasets?data_catalog_id={kielipankki_datacatalog_id}&persistent_identifier={dataset_pid}"
     assert result
+
 
 def test_check_if_dataset_record_pid_not_in_datacatalog(mock_requests_get_record, metax_base_url, kielipankki_datacatalog_id):
     """Test that a nn-existing PID in Metax is handled as expected."""
@@ -64,18 +35,6 @@ def test_get_dataset_record_metax_id(dataset_pid, mock_requests_get_record):
     assert mock_requests_get_record.request_history[0].url == "https://metax-service.fd-staging.csc.fi/v3/datasets?data_catalog_id=urn:nbn:fi:att:data-catalog-kielipankki-v3&persistent_identifier=urn.fi/urn:nbn:fi:lb-2016101210"
 
 
-@pytest.fixture
-def mock_post_put_response_json():
-    """Mock an id of a dataset in Metax."""
-    return _get_file_as_string("tests/test_data/put_post_response.json")
-
-@pytest.fixture
-def mock_requests_post(mock_post_put_response_json, metax_base_url):
-    """Mock a post request to Metax."""
-    with requests_mock.Mocker() as mocker:
-        mocker.post(f"{metax_base_url}/datasets", text=mock_post_put_response_json)
-        yield mocker
-
 def test_create_dataset_successful(mock_requests_post, caplog):
     """Check that a successful post request to Metax is made of a well-formed dictionary."""
     metadata_dict = {"persistent_identifier": "urn.fi/urn:nbn:fi:lb-201603170300", "title": {"en": "The Corpus"}, "description": {"en": "A large corpus"}, "modified": "2016-03-17T00:00:00.000000Z", "issued": "2016-03-17T00:00:00.000000Z", "access_rights": {"license": [{"url": "http://uri.suomi.fi/codelist/fairdata/license/code/undernegotiation"}], "access_type": {"url": "http://uri.suomi.fi/codelist/fairdata/access_type/code/restricted"}}}
@@ -88,10 +47,10 @@ def test_create_dataset_successful(mock_requests_post, caplog):
     assert mock_requests_post.request_history[0].method == "POST"
     assert mock_requests_post.request_history[0].url == "https://metax-service.fd-staging.csc.fi/v3/datasets"
 
+
 def test_create_dataset_failed(mock_requests_post, caplog, metax_base_url):
     """Check that an ill-formed dictionary results in a bad request to Metax."""
     mock_requests_post.post(f"{metax_base_url}/datasets", status_code=400)
-
     metadata_dict = {"persistent_identifier": "urn.fi/urn:nbn:fi:lb-201603170300", "title": {"en": "The Corpus"}, "description": {"en": "A large corpus"}, "modified": "2016-03-17T00:00:00.000000Z", "issued": "2016-03-17T00:00:00.000000Z", "access_rights": {"license": [{"url": "http://uri.suomi.fi/codelist/fairdata/license/code/undernegotiation"}], "access_type": {"orl": "http://uri.suomi.fi/codelist/fairdata/access_type/code/restricted"}}}
 
     with pytest.raises(requests.exceptions.HTTPError), caplog.at_level(logging.ERROR):
@@ -101,14 +60,6 @@ def test_create_dataset_failed(mock_requests_post, caplog, metax_base_url):
     assert mock_requests_post.request_history[0].method == "POST"
     assert mock_requests_post.request_history[0].url == "https://metax-service.fd-staging.csc.fi/v3/datasets"
 
-
-@pytest.fixture
-def mock_requests_put(mock_post_put_response_json, metax_base_url):
-    """Mock a PUT request of a dataset to Metax."""
-    metax_dataset_id = "441560f5-4c2a-48eb-bc1a-489639ec3573"
-    with requests_mock.Mocker() as mocker:
-        mocker.put(f"{metax_base_url}/datasets/{metax_dataset_id}", text=mock_post_put_response_json)
-        yield mocker
 
 def test_update_dataset_successful(mock_requests_put, caplog):
     """Test that an existing dataset in Metax is successfully updated."""
@@ -136,4 +87,3 @@ def test_update_dataset_failed(mock_requests_put, caplog, metax_base_url):
     assert len(mock_requests_put.request_history) == 1
     assert mock_requests_put.request_history[0].method == "PUT"
     assert mock_requests_put.request_history[0].url == "https://metax-service.fd-staging.csc.fi/v3/datasets/441560f5-4c2a-48eb-bc1a-489639ec3573"
-
